Fix super annoying property bug
SQLA was super angry that I was passing a <property> object rather than
a string when doing a get_or_create for a PropertyNamespace. I spent
a unreasonable amount of time trying to figure out why the property `namespace`
would be a <Property> object (as in the SQLA model). Turns out I was just using
the class rather than the object, and since `namespace` is a @property I got
that error. Switching to the object solved it directly.

I love you Python, but this was just weird. It was completely my fault, but
wth. :D